Rename WS messages namespace
