+++ vATIS.Desktop/Ui/Services/Websocket/Messages/AtisStationMessage.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Text.Json.Serialization;
 using Vatsim.Vatis.Profiles.Models;
 
-namespace Vatsim.Vatis.Ui.Services.Websocket.WebsocketMessages;
+namespace Vatsim.Vatis.Ui.Services.Websocket.Messages;
 
 /// <summary>
 /// Represents a message sent over the websocket with a list of ATIS stations.
