+++ vATIS.Desktop/Ui/Services/Websocket/Messages/LoadProfileMessage.cs
@@ -5,7 +5,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Vatsim.Vatis.Ui.Services.Websocket.WebsocketMessages;
+namespace Vatsim.Vatis.Ui.Services.Websocket.Messages;
 
 /// <summary>
 /// Represents a message sent over the websocket to change the active profile.
